+++ packages/components/src/breadcrumbs/index.ts
@@ -1,2 +1,2 @@
 export * from './Breadcrumbs'
-export type { BreadcrumbItem, BreadcrumbsProps } from './Breadcrumbs'
+export * from './Breadcrumb'